fitp-email-agent/requirements.txt
requests==2.33.1
python-dotenv==1.2.2
geopy==2.4.1
beautifulsoup4==4.14.3
lxml==6.0.4
